Prevents duplicate help messages being printed out, when using ./pt -h
The err message itself contains the entirety of the help message, plus
the flags library has already output, so when catching and displaying
the error, it duplicates the help message.

The check simply looks for 'Usage:' which is an indicator that it is a
help message and not an actual error.