Print error rather than exit when encountering permission denied
Previously if we encountered a file we could not read for permission
reasons, we'd exit immediately. This meant that you either had to stop
and fix the permissions or tell pt to ignore the file.

Now we print the error out which is similar to what grep does.

Rather than using log.Printf which could have issues writing at the same
time as printing the matches, this uses the same printer. It also adds a
new colour code for errors so that they are more noticable when running
with coloured output.

This should resolve #145.